json.dumps does not guarantee return unicode, wrap it with unicode()
json.dumps(data, ensure_ascii=False)
According to python doc, "If ensure_ascii is False, then the return value will be a unicode instance.".
But, if data does not contain unicode string, it will return a str. which in bypy jsondump caused an error:
TypeError: must be unicode, not str